chore(simulator): cleanup
