+++ crates/simulate/src/lib.rs
@@ -2,6 +2,8 @@
 
 mod core;
 mod interface;
+mod utils;
 
 pub use core::*;
 pub use interface::*;
+pub(crate) use utils::*;
