+++ crates/simulate/src/utils/mod.rs
@@ -1,0 +1,4 @@
+/// event utils
+pub(crate) mod event;
+/// evm utilities
+pub(crate) mod evm;
